requirements.txt
# Sprint 2: AI Plagiarism Detection - Complete Production Dependencies

# Core AI/ML
streamlit==1.28.1
sentence-transformers==2.2.2
huggingface_hub==0.25.2
torch>=2.0.1
torchvision>=0.15.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==1.5.3
tqdm==4.66.1

# Vector Database (FAISS with fallback)
# FAISS - may fail on Windows, scikit-learn fallback available
faiss-cpu==1.7.4

# FastAPI Backend (NEW for Sprint 2)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0

# Document Processing
pdfplumber==0.9.0
python-docx==0.8.11

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Security & Utilities
filelock==3.13.1
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0

# Optional: MongoDB (for production)
pymongo==4.6.0
